+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- **breaking:** replaced `OwnedSlice`'s `owned_slice_ptr` method with `owned_slice_ref` which returns a reference instead of a pointer
+
 ## 0.15.2 (2025-01-06)
 - **fixed:** `append` implementation of `MutBumpVecRev` now accounts for `take_owned_slice` panicking
 - **added:** implemented `Default` for `Stats`, `ChunkPrevIter`, `ChunkNextIter`, `WithoutShrink` and `WithoutDealloc`