+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.16.0 (2025-01-10)
 - **breaking:** replaced `OwnedSlice`'s `owned_slice_ptr` method with `owned_slice_ref` which returns a reference instead of a pointer
 - **fixed:** `Write` impl for `FixedBumpVec<u8>` not actually writing
 - **added:** `unsize_bump_box` to unsize `BumpBox`es on stable